+++ wallet/api/request_permissions_test.go
@@ -30,6 +30,7 @@ func TestRequestPermissions(t *testing.T) {
 	t.Run("Using wrong passphrase doesn't update the permissions", testUsingWrongPassphraseDoesNotUpdatePermissions)
 	t.Run("Getting internal error during the wallet retrieval doesn't update the permissions", testGettingInternalErrorDuringWalletRetrievalDoesNotUpdatePermissions)
 	t.Run("Getting internal error during the wallet saving doesn't update the permissions", testGettingInternalErrorDuringWalletSavingDoesNotUpdatePermissions)
+	t.Run("Updating the permissions doesn't overwrite untracked changes", testUpdatingPermissionsDoesNotOverwriteUntrackedChanges)
 }
 
 func testRequestingPermissionsWithInvalidParamsFails(t *testing.T) {
@@ -635,8 +636,19 @@ func testGettingInternalErrorDuringWalletSavingDoesNotUpdatePermissions(t *testi
 	// given
 	ctx, traceID := contextWithTraceID()
 	hostname := "vega.xyz"
+	walletName := vgrand.RandomStr(5)
 	originalPermissions := wallet.Permissions{}
-	wallet1 := walletWithPerms(t, hostname, originalPermissions)
+	wallet1, recoveryPhrase, err := wallet.NewHDWallet(walletName)
+	if err != nil {
+		t.Fatal("couldn't create wallet for test: %w", err)
+	}
+
+	// Clone the wallet1, so we can emulate a different instance returned by
+	// the wallet store.
+	loadedWallet, err := wallet.ImportHDWallet(walletName, recoveryPhrase, 2)
+	if err != nil {
+		t.Fatal("couldn't import wallet for test: %w", err)
+	}
 	passphrase := vgrand.RandomStr(5)
 	requestedPermissions := map[string]string{
 		"public_keys": "read",
@@ -648,8 +660,8 @@ func testGettingInternalErrorDuringWalletSavingDoesNotUpdatePermissions(t *testi
 	// -- expected calls
 	handler.pipeline.EXPECT().RequestPermissionsReview(ctx, traceID, hostname, wallet1.Name(), requestedPermissions).Times(1).Return(true, nil)
 	handler.pipeline.EXPECT().RequestPassphrase(ctx, traceID, wallet1.Name()).Times(1).Return(passphrase, nil)
-	handler.walletStore.EXPECT().GetWallet(ctx, wallet1.Name(), passphrase).Times(1).Return(wallet1, nil)
-	handler.walletStore.EXPECT().SaveWallet(ctx, wallet1, passphrase).Times(1).Return(assert.AnError)
+	handler.walletStore.EXPECT().GetWallet(ctx, wallet1.Name(), passphrase).Times(1).Return(loadedWallet, nil)
+	handler.walletStore.EXPECT().SaveWallet(ctx, loadedWallet, passphrase).Times(1).Return(assert.AnError)
 	handler.pipeline.EXPECT().NotifyError(ctx, traceID, api.InternalError, fmt.Errorf("couldn't save wallet: %w", assert.AnError)).Times(1)
 	// -- unexpected calls
 	handler.walletStore.EXPECT().WalletExists(gomock.Any(), gomock.Any()).Times(0)
@@ -674,6 +686,71 @@ func testGettingInternalErrorDuringWalletSavingDoesNotUpdatePermissions(t *testi
 	connectedWallet, err := handler.sessions.GetConnectedWallet(token)
 	require.NoError(t, err)
 	assert.Equal(t, originalPermissions.Summary(), connectedWallet.Permissions().Summary())
+}
+
+func testUpdatingPermissionsDoesNotOverwriteUntrackedChanges(t *testing.T) {
+	// given
+	ctx, traceID := contextWithTraceID()
+	hostname := "vega.xyz"
+	walletName := vgrand.RandomStr(5)
+	wallet1, recoveryPhrase, err := wallet.NewHDWallet(walletName)
+	if err != nil {
+		t.Fatal("couldn't create wallet for test: %w", err)
+	}
+
+	// Clone the wallet1, so we can modify the clone without tempering with wallet1.
+	modifiedWallet, err := wallet.ImportHDWallet(walletName, recoveryPhrase, 2)
+	if err != nil {
+		t.Fatal("couldn't import wallet for test: %w", err)
+	}
+	kp, _ := modifiedWallet.GenerateKeyPair([]wallet.Meta{{Key: "name", Value: "hello"}})
+
+	passphrase := vgrand.RandomStr(5)
+	askedPermissions := wallet.PermissionsSummary{
+		"public_keys": "read",
+	}
+
+	// setup
+	handler := newRequestPermissionsHandler(t)
+	token := connectWallet(t, handler.sessions, hostname, wallet1)
+	// -- expected calls
+	handler.pipeline.EXPECT().RequestPermissionsReview(ctx, traceID, hostname, wallet1.Name(), askedPermissions).Times(1).Return(true, nil)
+	handler.pipeline.EXPECT().RequestPassphrase(ctx, traceID, wallet1.Name()).Times(1).Return(passphrase, nil)
+	handler.walletStore.EXPECT().GetWallet(ctx, wallet1.Name(), passphrase).Times(1).Return(modifiedWallet, nil)
+	handler.walletStore.EXPECT().SaveWallet(ctx, gomock.Any(), passphrase).Times(1).DoAndReturn(func(_ context.Context, w wallet.Wallet, _ string) error {
+		// We verify that the saved wallet contains the modification from the modified wallet and the permissions update from wallet1.
+		assert.Equal(t, []wallet.KeyPair{kp}, w.ListKeyPairs())
+		assert.Equal(t, wallet.Permissions{
+			PublicKeys: wallet.PublicKeysPermission{
+				Access: wallet.ReadAccess,
+			},
+		}, w.Permissions(hostname))
+		return nil
+	})
+	handler.pipeline.EXPECT().NotifySuccessfulRequest(ctx, traceID).Times(1)
+	// -- unexpected calls
+	handler.walletStore.EXPECT().WalletExists(gomock.Any(), gomock.Any()).Times(0)
+	handler.walletStore.EXPECT().ListWallets(gomock.Any()).Times(0)
+	handler.pipeline.EXPECT().RequestWalletConnectionReview(gomock.Any(), gomock.Any(), gomock.Any()).Times(0)
+	handler.pipeline.EXPECT().RequestWalletSelection(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(0)
+	handler.pipeline.EXPECT().RequestTransactionReview(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(0)
+	handler.pipeline.EXPECT().NotifyError(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(0)
+	handler.pipeline.EXPECT().NotifyTransactionStatus(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Times(0)
+
+	// when
+	result, errorDetails := handler.handle(t, ctx, api.RequestPermissionsParams{
+		Token:                token,
+		RequestedPermissions: askedPermissions,
+	})
+
+	// then
+	assert.Nil(t, errorDetails)
+	require.NotEmpty(t, result)
+	assert.Equal(t, askedPermissions, result.Permissions)
+	// Verifying the connected wallet is updated.
+	connectedWallet, err := handler.sessions.GetConnectedWallet(token)
+	require.NoError(t, err)
+	assert.Equal(t, askedPermissions, connectedWallet.Permissions().Summary())
 }
 
 type requestPermissionsHandler struct {